requirements.txt
fastapi
uvicorn 
typing
pydantic
uvicorn 
sqlalchemy
passlib
bcrypt
pydantic[email]
python-jose
datetime
Oauth2
python-multipart